ai-everyone/agents/teams-agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
requests>=2.32.0
msal>=1.31.0
python-dotenv>=1.0.0
pydantic>=2.10.0
